# src/api/docs/openapi.yaml
openapi: 3.0.1
info:
  title: EPS API
  description: >-
    This is the API which supports the web and mobile applications for the
    Extended Prototype System (EPS).
  contact:
    email: jclements@ascend.network
  version: 0.29.0
...
paths:
...
  /case_definitions:
    post:
      tags:
        - case_definitions
      summary: Add case definition to system.
      description: |
        This endpoint can be used to add new case definitions to the system. It
        accepts objects as described in the **Request body** section below. It
        also performs data validation on the submitted object as described in
        the **Responses** section.

        This endpoint is protected by JWT token submission in the
        Authorization header and the account that calls this endpoint must have
        the `CREATE_CASE_DEFINITION` permission.

        If the account has permission and a valid case definition object is
        submitted, the case definition will be added to the system and the
        endpoint will return a full case definition object as described in the
        **Responses** section.
      requestBody:
        description: |
          The endpoint accepts the following object to create new case
          definitions. The only property that is absolutely required is
          `name`. However, if you submit a survey ID in the `surveys` list it
          must be a system ID from a survey in the system. Also it you submit
          a case management document as part of the `documents` list, it must
          have a name.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCaseDefinitionObject'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDefinition'
        '400':
          description: |
            If invalid data is submitted by a client, a 400 status is returned
            with a contextual message. The endpoint validates incoming data
            for the following issues.
            * Missing or empty case definition name
            * Duplicate definition name
            * An unknown survey ID
            * Missing or empty case definition document name
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: [ ]
    get:
      tags:
        - case_definitions
      summary: Get list of case definitions
      description: |
        Returns array of all case definitions in the system. 
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfCaseDefinitions'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: [ ]
  /case_definitions/<int:case_definition_id>:
    get:
      tags:
        - case_definitions
      summary: Get a case definition
      description: |
        Returns a case definition based on the `case_definition_id` in the URL.
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDefinition'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - case_definitions
      summary: Update a case definition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseDefinition'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDefinition'
    delete:
      tags:
        - case_definitions
      summary: Delete a case definition
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
  /case_definitions/<int:case_definition_id>/custom_fields:
    post:
      tags:
        - case_definitions
      summary: Add custom field to case definition.
      parameters:
        - in: path
          name: case_definition_id
          schema:
            type: integer
          required: true
          description: The case definition ID of the custom field being updated.
          example: 16
      description: |
        This endpoint can be used to add a custom field to a case definition.

        It will also add the custom field to all cases of the case definition.

        Accounts need the UPDATE_CASE_DEFINITION permission.
      requestBody:
        description: |
          The endpoint accepts the following object to create new cases.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCustomFieldObject'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomField'
        '400':
          description: |
            If invalid data is submitted by a client, a 400 status is returned
            with a contextual message. The endpoint validates incoming data
            for the following issues.
            * Missing or empty custom field name
            * Missing or empty custom field field type
            * A selection object without an id property
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: A 404 is returned if an unknown case definition is used.
      security:
        - bearerAuth: [ ]
    get:
      tags:
        - case_definitions
      summary: Get all custom fields for case definition.
      parameters:
        - in: path
          name: case_definition_id
          schema:
            type: integer
          required: true
          description: The case definition ID of the custom field being updated.
          example: 16
      description: |
        This endpoint returns a list of all custom fields for a case
        definition.

        Accounts need the READ_CASE_DEFINITION permission.
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfCustomFields'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: A 404 is returned if an unknown case definition is used.
      security:
        - bearerAuth: [ ]
  /case_definitions/<int:case_definition_id>/custom_fields/<custom_field_id>:
    get:
      tags:
        - case_definitions
      summary: Get a case definition's custom field
      parameters:
        - in: path
          name: case_definition_id
          schema:
            type: integer
          required: true
          description: The case definition ID of the custom field being updated.
          example: 16
        - in: path
          name: custom_field_id
          schema:
            type: integer
          required: true
          description: The custom field's ID.
          example: 11
      description: |
        Returns the custom field for the case definition ID and custom field ID in the URL.

        Users needs the READ_CASE_DEFINITION permission to use this endpoint.
      responses:
        '200':
          description: The custom field's value was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomField'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A 404 is returned if a unknown case or custom field is requested.
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - case_definitions
      summary: Update a case definition's custom field
      parameters:
        - in: path
          name: case_definition_id
          schema:
            type: integer
          required: true
          description: The case definition ID of the custom field being updated.
          example: 16
        - in: path
          name: custom_field_id
          schema:
            type: integer
          required: true
          description: The custom field's ID.
          example: 11
      description: |
        This endpoint can be used to update a case definition's custom field.

        It will also update the custom field in all cases of the case definition.

        Accounts need UPDATE_CASE_DEFINITION permission
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomFieldObject'
        required: true
      responses:
        '200':
          description: The custom field's value was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomField'
        '400':
          description: |
            If invalid data is submitted by a client, a 400 status is returned
            with a contextual message. The endpoint validates incoming data
            for the following issues.
            * Missing or empty custom field name
            * Missing or empty custom field field type
            * A selection object without an id property
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: |
            404 is returned if an unknown case definition or custom field ID is
            used in the URL.
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - case_definitions
      summary: Delete a case definition's custom field
      parameters:
        - in: path
          name: case_definition_id
          schema:
            type: integer
          required: true
          description: The case definition ID of the custom field being updated.
          example: 16
        - in: path
          name: custom_field_id
          schema:
            type: integer
          required: true
          description: The custom field's ID.
          example: 11
      description: |
        Deletes the custom field for the case definition ID and custom field ID in the URL.

        It will also delete the custom field from all cases of the case definition.

        Users needs the UPDATE_CASE_DEFINITION permission to use this endpoint.
      responses:
        '200':
          description: The custom field's was successfully deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A 404 is returned if a unknown case definition or custom field is requested.
      security:
        - bearerAuth: [ ]
  /case_definitions/<int:case_definition_id>/cases:
    post:
      tags:
        - case_definitions
      summary: Add case the to system.
      description: |
        This endpoint can be used to add new cases to the system. It
        accepts objects as described in the **Request body** section below. It
        also performs data validation on the submitted object as described in
        the **Responses** section.

        This endpoint is protected by JWT token submission in the
        Authorization header and the account that calls this endpoint must have
        the `CREATE_CASE` permission.

        If the account has permission and a valid case object is submitted,
        the case will be added to the system and the endpoint will return a
        full case object as described in the **Responses** section.

        Also on successful submission, the system will generate a `key` for the
        case. The `key` follows the format:
        `<case_definition_key>-<next_sequential_case_number>` where 
        `case_definition_key` is the `key` for the Case Definition this Case is
        related to and `next_sequential_case_number` is next sequential number
        for Cases related to this Case Definition. For example, if the 
        Case Definitions key is CD and there are 10 cases associated with this
        case definition, then the next case's key will be CD-11.

        The endpoint will accept a `assigned_to_id` inwhich to use to assign a
        user to the case. If the ID is invalid or unknown the endpoint will
        return an error. In an assigned to is send, the endpoint will
        record the timestamp of the assignment. The user referenced by the ID
        must have the `ASSIGNABLE_TO_CASE` permission or the API will return
        an error.

        This endpoint also accepts location data. If location data is sent,
        it will be logged as the location the case was created.

        The endpoint accepts a list of custom fields for the case. If a
        `custom_fields` property is found and that property is a list, the
        endpoint will, assuming the list is valid, use that list as the list of
        custom fields for the case. The endpoint will validate the custom
        fields sent ensuring the following:

          * That each submitted custom field matches a custom field for the case's case definition
          * That the submitted custom fields have a unique ID
          * That the case added to the system has all the custom fields that it's case defintion has. Any of the case's case definition's custom fields that are not submitted are added.

        The endpoint will also created and updated values for each custom field.
        If `create_at`, `created_by_id`, `updated_at`, `updated_by_id` are sent,
        those values will be used instead of generated by the system.
      requestBody:
        description: |
          The endpoint accepts the following object to create new cases. The
          property that is required is `name`.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCaseObject'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          description: |
            If invalid data is submitted by a client, a 400 status is returned
            with a contextual message. The endpoint validates incoming data
            for the following issues.
            * Missing or empty case name
            * Duplicate case name
            * Case names that are too short or too long.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: A 404 is returned if a unknown case definition is used.
      security:
        - bearerAuth: [ ]
  /cases:
    get:
      tags:
        - cases
      summary: Get list of cases
      description: |
        Returns list of cases in the system. The account used to make this
        call must have permission to read cases. The endpoint can perform
        case searches using the `search_term` parameter defined below.
      parameters:
        - in: query
          name: search_term
          schema:
            type: string
          description: |
            This query parameter can be used to search cases. The search term
            passed in will be used to search the case's name and key. The
            search is case insensitive. The search term is not required and if
            not used or if it is empty (i.e. `/cases?search_term=`) the endpoint
            will return all cases in the system.
        - in: query
          name: case_definition_id
          schema:
            type: number
            description: |
              Using this query parameter will return only the cases for the case
              definition ID passed.
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfCases'
        '401':
          description: Authorization information is missing or invalid.
      security:
        - bearerAuth: [ ]
  /cases/<case_id>:
    get:
      tags:
        - cases
      summary: Get a case
      description: |
        Returns the case object in JSON for the `<case_id>` that is part of
        the URL.  The account used to make this call must have permission to
        read cases.
      responses:
        '200':
          description: success operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A 404 is returned if a unknown case is requested.
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - cases
      summary: Update a case
      description: |
        This endpoint can be used to update a case.  It
        accepts objects as described in the **Request body** section below. It
        also performs data validation on the submitted object as described in
        the **Responses** section.

        This endpoint is protected by JWT token submission in the
        Authorization header and the account that calls this endpoint must
        have permission to update cases.

        The endpoint will accept a `assigned_to_id` inwhich to use to assign a
        user to the case. If the ID is invalid or unknown the endpoint will
        return an error. In an assigned to is send, the endpoint will
        record the timestamp of the assignment.  The user referenced by the ID
        must have the `ASSIGNABLE_TO_CASE` permission or the API will return
        an error.

        This endpoint also accepts location data. If location data is sent,
        it will be logged as the location the case was updated.
      requestBody:
        description: |
          The endpoint accepts the following object to update cases, any other
          properties will be ignored. Neither property is required, however
          if a `name` property is submitted it must:
            * Cannot be null or empty
            * Be between 8 and 50 characters
            * Not be the same as another case

          If `status_id` is not provided, it will not be updated. The
          `status_id` is validated as a valid status id number.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseObject'
        required: true
      responses:
        '200':
          description: success operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          description: |
            If invalid data is submitted by a client, a 400 status is returned
            with a contextual message. The endpoint validates incoming data
            for the following issues.
              * If a name property exists:
                * It cannot be an empty string or null
                * It cannot be duplicate case name
                * It must be between 8-50 characters long
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A 404 is returned if a unknown case is requested.
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - cases
      summary: Delete a case
      description: |
        This endpoint can be used to delete a case. It will delete the case
        referenced by the `<case_id>` in the URL.

        This endpoint is protected by JWT token submission in the
        Authorization header and the account that calls this endpoint must
        have permission to update cases.
      responses:
        '200':
          description: Success Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A 404 is returned if a unknown case is requested.
      security:
        - bearerAuth: [ ]
  /cases/<case_id>/custom_fields:
    get:
      tags:
        - cases
      summary: Get all case's custom fields
      parameters:
        - in: path
          name: case_id
          schema:
            type: integer
          required: true
          description: The case ID of the custom field being updated.
          example: 16
      description: |
        Returns the custom fields for the case ID in the URL.

        Users needs the READ_CASE permission to use this endpoint.
      responses:
        '200':
          description: The custom field's value was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfCaseCustomFields'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A 404 is returned if a unknown case is requested.
      security:
        - bearerAuth: [ ]
  /cases/<case_id>/custom_fields/<custom_field_id>:
    get:
      tags:
        - cases
      summary: Get a case's custom field
      parameters:
        - in: path
          name: case_id
          schema:
            type: integer
          required: true
          description: The case ID of the custom field being updated.
          example: 16
        - in: path
          name: custom_field_id
          schema:
            type: string
          required: true
          description: The custom field's ID.
          example: c7b8dcf3-c9fb-4d60-9c9f-6d7016155619
      description: |
        Returns the custom field for the case ID and custom field ID in the URL.

        Users needs the READ_CASE permission to use this endpoint.
      responses:
        '200':
          description: The custom field's value was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseCustomField'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A 404 is returned if a unknown case or custom field is requested.
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - cases
      summary: Update the value of a case's custom field
      parameters:
        - in: path
          name: case_id
          schema:
            type: integer
          required: true
          description: The case ID of the custom field being updated.
          example: 16
        - in: path
          name: custom_field_id
          schema:
            type: string
          required: true
          description: The custom field's ID.
          example: c7b8dcf3-c9fb-4d60-9c9f-6d7016155619
      description: |
        This endpoint can be used to update the value of a case's custom
        field.
      requestBody:
        description: |
          The endpoint accepts an object with a `value` property. The `value`
          property is not required but if it is sent, the value of the custom
          field will be updated to whatever the value of the property is. If
          `null` or an empty string is sent, the value of the custom field
          will be updated to that value. If the request body does not contain
          a `value` property this endpoint will do nothing.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomFieldValueObject'
        required: true
      responses:
        '200':
          description: The custom field's value was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseCustomField'
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: |
            404 is returned if an unknown case or custom field ID is used in
            the URL.
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    AddCaseDefinitionObject:
      type: object
      properties:
        name:
          type: string
        key:
          type: string
        description:
          type: string
        surveys:
          type: array
          description: |
            Array of the system IDs for the surveys associated with the case
            definition.
          items:
            type: integer
        documents:
          type: array
          items:
            $ref: '#/components/schemas/AddCaseDefinitionDocumentObject'
        custom_fields:
          type: array
          description: List of custom fields for this case definition.
          items:
            $ref: '#/components/schemas/AddCustomFieldObject'
        activity_definitions:
          type: array
          items:
            $ref: '#/components/schemas/AddActivityDefinitionObject'
      required:
        - name
        - key
...
    CaseDefinition:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        key:
          type: string
        description:
          type: string
        surveys:
          type: array
          items:
            $ref: '#/components/schemas/CaseDefinitionSurveyObject'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/AddCaseDefinitionDocumentObject'
        custom_fields:
          type: array
          items:
            $ref: '#/components/schemas/CustomField'
        activity_definitions:
          type: array
          items:
            $ref: '#/components/schemas/ActivityDefinition'
        created_at:
          type: string
        updated_at:
          type: string
        created_by:
          # type: object
          $ref: '#/components/schemas/ActionByUser'
        updated_by:
          # type: object
          $ref: '#/components/schemas/ActionByUser'
...
    ArrayOfCaseDefinitions:
      type: array
      items:
        $ref: '#/components/schemas/CaseDefinition'
...
    ArrayOfCases:
      type: array
      items:
        $ref: '#/components/schemas/Case'
...
    ContextualMessage:
      type: object
      properties:
        message:
          type: string
          description: A contextual error message.
...
    ArrayOfCaseCustomFields:
      type: array
      items:
        $ref: '#/components/schemas/CaseCustomField'
    ArrayOfCustomFields:
      type: array
      items:
        $ref: '#/components/schemas/CustomField'
...
  responses:
...
    UnauthorizedError:
      description: Unauthorized
    NotFoundError:
      description: NotFound
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT